% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColorPalette.R
\docType{data}
\name{ColorPalette}
\alias{ColorPalette}
\title{Color Palette for plotting}
\format{
A data frame with 9 rows and 2 columns:
\describe{
\item{Color}{Color name}
\item{Hex}{Hexadecimal code}
...
}
}
\source{
Color schemes were taken from \href{https://personal.sron.nl/~pault/}{Paul Tol's website.}
}
\usage{
ColorPalette
}
\description{
Color Palette I like to use for my plots.
}
\keyword{datasets}
