% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationMatchType.R
\name{correlationMatchType}
\alias{correlationMatchType}
\title{Generates match type plots for SILAC output from Fragpipe}
\usage{
correlationMatchType(df_digly, df_input = NULL)
}
\arguments{
\item{df_digly}{Dataframe corresponding to combine_modified_peptide_label_quant.tsv for the digly data.}

\item{df_input}{Dataframe corresponding to combine_modified_peptide_label_quant.tsv for the input data. Default is NULL.}
}
\value{
List of dataframes with calculated match type percentages and plots for visualization.
}
\description{
This function takes a dataframe from the Fragpipe output combine_modified_peptide_label_quant.tsv,
calculates the percentage of pairs that were matched either by MS/MS or requantify and generates
plots to visualize the percentage of matches.
}
