% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/danRer11_106GeneLengths.R
\docType{data}
\name{danRer11_106GeneLengths}
\alias{danRer11_106GeneLengths}
\title{Gene lengths from Ensembl danRer11 106}
\format{
A data frame with 32520 rows and 2 columns:
\describe{
\item{ensembl_gene_id}{Ensembl Gene ID}
\item{Length}{Longest transcript length}
...
}
}
\source{
Generated with biomaRt.
}
\usage{
danRer11_106GeneLengths
}
\description{
Dataframe with gene lengths for zebrafish genome from Ensemble danRer11 106 for use in edgeR.
}
\keyword{datasets}
