% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/danRer11GeneNames.R
\docType{data}
\name{danRer11GeneNames}
\alias{danRer11GeneNames}
\title{Mapped gene names to Ensemble Gene and Transcript IDs}
\format{
A data frame with 59876 rows and 3 columns:
\describe{
\item{ensembl_gene_id}{Ensembl Gene ID}
\item{ensembl_transcript_id}{Ensembl Transcript ID}
\item{Name}{Gene Name}
...
}
}
\source{
Generated with biomaRt.
}
\usage{
danRer11GeneNames
}
\description{
Dataframe with mapped gene names to Ensemble Gene and Transcript IDs for danRer11.
}
\keyword{datasets}
