% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diannQCreport.R
\name{diannQCreport}
\alias{diannQCreport}
\title{Generates QC reports from DIA data from DIA-NN}
\usage{
diannQCreport(inputPath, jobname, outputPath)
}
\arguments{
\item{inputPath}{Input path where "metadata.csv", "report.tsv", "report.stats.tsv" and the fasta file are.}

\item{jobname}{Specifies the folder name to save .}

\item{outputpath}{Output path to save all the outputs.}
}
\value{
Generates the QC reports for each run and saves them in the outputPath.
}
\description{
This function takes dia-nn output and generate QC reports that were missing from
fragpipe using dia-nn v.2.2.0. All reports are based on each raw file run.
}
