% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DigestBatch.R
\name{DigestBatch}
\alias{DigestBatch}
\title{Digest multiple protein sequences}
\usage{
DigestBatch(df, enz = "trypsin", miss = 1)
}
\arguments{
\item{df}{Dataframe input with aminoacid sequence in a columm named 'Sequence', protein ids in a column names 'Entry' and gene names in a column named 'Gene Names'}

\item{enz}{Specifies enzyme to be used. Options are "trypsin" (default), "trypsin.strict", "pepsin"}

\item{miss}{Missed cleavages to account for. Default is 1.}
}
\value{
Generates a dataframe with all predicted digested peptides form all protein sequences.
}
\description{
This function takes a dataframe downloaded from UNIPROT with Entry (Protein ID), Gene Names
and Sequence columns, then uses Digest from OrgMassSpecR and loops for multiple sequences
in the dataframe. Output is a dataframe with all digested peptides.
}
