% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E3LigaseKK.R
\docType{data}
\name{E3LigaseKK}
\alias{E3LigaseKK}
\title{E3 Ligase list from Kostova Lab}
\format{
A data frame with 377 rows and 1 columns:
\describe{
\item{protein_name}{Gene names}
...
}
}
\source{
Sent by Kamena Kostova on 2025.
}
\usage{
E3LigaseKK
}
\description{
Dataframe with E3 ligases from Kostova Lab.
}
\keyword{datasets}
