% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/essentialTranslationZf.R
\docType{data}
\name{essentialTranslationZf}
\alias{essentialTranslationZf}
\title{Essential translation factors by Leo (Kostova Lab)}
\format{
A data frame:
\describe{
\item{gene_name}{Gene name by Leo}
\item{protein_name}{Protein name}
\item{protein_id}{Protein ID}
\item{description}{Leo's description of function}
...
}
}
\source{
Got from Leo in Kostova Lab.
}
\usage{
essentialTranslationZf
}
\description{
Dataframe with essential translation factors.
}
\keyword{datasets}
