% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastaTransformUniprot.R
\name{fastaTransform}
\alias{fastaTransform}
\title{Transforms a dataframe with protein sequences into FASTA format}
\usage{
fastaTransform(df, organism = "DanRer", outputpath)
}
\arguments{
\item{df}{Dataframe input with aminoacid sequence in a columm named 'Sequence' and protein ids in a column names 'Entry'}

\item{organism}{Specifies the organism the proteins are from. Default is 'DanRer'.}

\item{outputpath}{Output path to save the final FASTA file.}
}
\value{
Generates a FASTA file based on a dataframe with aminoacid sequences.
}
\description{
This function takes a dataframe downloaded from UNIPROT with Entry (Protein ID)
and Sequence columns, then outputs a FASTA file.
}
