% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folderStructure.R
\name{folderStructure}
\alias{folderStructure}
\title{Creates folder structure for PROT-XXXX projects}
\usage{
folderStructure(path)
}
\arguments{
\item{path}{Path to your PROT-XXXX folder}
}
\value{
Creates directories inside PROT-XXXX folder
}
\description{
This function creates the folder structure inside the path you input that is used for all other
mass spec related functions from gsptools package.
}
