% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromList.R
\name{fromList}
\alias{fromList}
\title{Gets names of genes from an upsetR object}
\usage{
fromList(input)
}
\arguments{
\item{input}{upsetR object}
}
\value{
Generates a dataframe with your gene names per group
}
\description{
This function takes an upsetR object and gives you the names for each intersection group
This was created by: https://github.com/docmanny
}
