% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateMsdapScript.R
\name{generateMsdapScript}
\alias{generateMsdapScript}
\title{Generates msdap R script}
\usage{
generateMsdapScript(inputPath, outputPath)
}
\arguments{
\item{inputPath}{Path where input files report.tsv (diann), sample_metadata.xlsx, FASTA used in the search and contrast.csv are located.}

\item{outputPath}{Path where msdap results will be saved.}
}
\value{
Creates an R script to run msdap.
}
\description{
This function generates an R script to run msdap.
It is a helper function for the processDiannMSdap.R function.
}
\details{
Settings for MSdap are the defaults for DIA data, if necessary change inside the first
glue::glue block.

NOTE: sample_metadata.xlsx can be generated manually or with the function
write_template_for_sample_metadata() from library(msdap).
}
