% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatePostprocessingScript.R
\name{generatePostprocessingScript}
\alias{generatePostprocessingScript}
\title{Generates post-processing R script}
\usage{
generatePostprocessingScript(outputPath)
}
\arguments{
\item{outputPath}{Path where results will be saved.}
}
\value{
Creates an R script to run post-processing analysis.
}
\description{
This function generates an R script to run post-processing on msdap output.
It is a helper function for the processDiannMSdap.R function.
}
