% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneNames_KEGG.R
\name{getGeneNames_KEGG}
\alias{getGeneNames_KEGG}
\title{Gets gene names from a KEGG pathway ID}
\usage{
getGeneNames_KEGG(pathwayID)
}
\arguments{
\item{pathwayID}{pathwayID should be either number or with organism specified. e.g., dre03010 where "dre" refers to Danio rerio}
}
\value{
Generates a vector with all your gene name associated to a KEGG pathway ID.
}
\description{
This function takes a KEGG pathway ID and outputs the gene names associated with them. Requires KEGGREST library.
}
