% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importDiannMSdap.R
\name{importDiannMSdap}
\alias{importDiannMSdap}
\title{Import diaNN output into MSdap for processing}
\usage{
importDiannMSdap(inputPath, outputPath)
}
\arguments{
\item{inputPath}{Path where the 4 files necessary are located}

\item{outputPath}{Path to where you want to save the MSdap output}
}
\value{
Creates 2 scripts to run MSdap on HPC.
}
\description{
This function takes report.tsv from diaNN, sample_metadata.xlsx from MSdap, a contrasts.csv file
and the FASTA file used in the proteomics search and creates an .R script to run MSdap and
a SLURM script to run in HPC cluster.
}
\details{
Settings for MSdap are the defaults for DIA data, if necessary change inside the first
glue::glue block.

NOTE: sample_metadata.xlsx can be generated manually or with the function
write_template_for_sample_metadata() from library(msdap).
}
