% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importFragpipeSILACdiGly.R
\name{importFragpipeSILACdiGly}
\alias{importFragpipeSILACdiGly}
\title{Imports SILAC diGly data from Fragpipe and performs analysis.}
\usage{
importFragpipeSILACdiGly(
  inputPath,
  psmDiglyPath,
  psmInputPath,
  jobname,
  outputPath,
  replicateFilter = FALSE,
  proteinInput = TRUE,
  groups = FALSE,
  force = FALSE,
  skipChromatogram = FALSE,
  exclude = NULL
)
}
\arguments{
\item{inputPath}{Input path where "metadata.csv", "contrast.csv", "digly_combined_modified_peptide_label_quant.tsv" and "input_combined_modified_peptide_label_quant.tsv" are located.}

\item{psmDiglyPath}{Path to where all psm.tsv files are for the digly search.}

\item{psmInputPath}{Path to where all psm.tsv files are for the input search.}

\item{jobname}{Specifies the folder name for convenience I always use the PROT-XXXX.}

\item{replicateFilter}{Defaults to FALSE. Filters proteins that are not present in at least 2 replicates. Since majority of SILAC experiments only have 2 replicates (swaps) this is FALSE by default.}

\item{proteinInput}{Defaults to TRUE. Includes protein input for comparisons.}

\item{groups}{Defaults to FALSE. If TRUE, will use the type column in metadata.csv to search unique proteins based on those groupings.}

\item{force}{Defaults to FALSE. Forces saving the results table in case it already exists.}

\item{skipChromatogram}{Defaults to FALSE. Skips generation of chromatogram plots.}

\item{exclude}{Defaults to NULL. Defines a character vector of columns to be excluded from the pipeline.}

\item{outputpath}{Output path to save all the outputs.}
}
\value{
Generates the results files and saves them in the outputPath.
}
\description{
This function takes an inputPath with "metadata.csv" and "combined_modified_peptide_label_quant.tsv"
generated, imports search results, combines site localization, perform t.test comparisons based on replicates
and outputs result files as an excel table for later import in ShinyApp.
}
