% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importFragpipeTMT.R
\name{importFragpipeTMT}
\alias{importFragpipeTMT}
\title{Imports TMT data from fragpipe and normalizes with NormalyzerDE}
\usage{
importFragpipeTMT(inputPath, jobname, outputPath, force = FALSE)
}
\arguments{
\item{inputPath}{Input path where "metadata.csv", "contrast.csv" and "abundance_protein_MD.tsv" are located.}

\item{jobname}{Specifies the folder name for NormalyzerDE, for convenience I always use the PROT-XXXX.}

\item{force}{Default to FALSE. Forces normalyzer DE to run.}

\item{outputpath}{Output path to save all the outputs.}
}
\value{
Generates the NormalyzerDE report from fragpipe input.
}
\description{
This function takes an inputPath with "metadata.csv", "contrast.csv" and
"abundance_protein_MD.tsv", performs normalization with NormalyzerDE and saves
all outputs in the ouputPath. Examples of input files can be found in inst/extdata/importFragpipeTMT_examples.
}
