% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaFragpipeTMT.R
\name{limmaFragpipeTMT}
\alias{limmaFragpipeTMT}
\title{Imports normalized TMT data from NormalyzerDE and performs limma contrasts.}
\usage{
limmaFragpipeTMT(
  inputPath,
  jobname,
  method = "quantile",
  outputPath,
  replicateFilter = TRUE,
  groups = FALSE,
  force = FALSE,
  exclude = NULL
)
}
\arguments{
\item{inputPath}{Input path where "metadata.csv", "contrast.csv" and "abundance_protein_MD.tsv" are located.}

\item{jobname}{Specifies the folder name for NormalyzerDE, for convenience I always use the PROT-XXXX.}

\item{method}{Specifies the chosen normalization method, defaults to "quantile". Options are: "cycloess", "gi", "log2", "mean", "median", "quantile", "rlr"}

\item{replicateFilter}{Defaults to TRUE. Filters proteins that are not present in at least 2 replicates.}

\item{groups}{Defaults to FALSE. If TRUE, will use the type column in metadata.csv to search unique proteins based on those groupings.}

\item{force}{Defaults to FALSE. Forces saving the results table in case it already exists.}

\item{exclude}{Defaults to NULL. Defines a character vector of columns to be excluded from the pipeline.}

\item{outputpath}{Output path to save all the outputs.}
}
\value{
Generates the limma contrasts and saves them in the outputPath.
}
\description{
This function takes an inputPath with "metadata.csv", "contrast.csv" and
"abundance_protein_MD.tsv", and the chosen normalization method from the
NormalizerDE output folder, then perform limma differential gene expression
and saves all outputs in the ouputPath. jobname should be the same used from
importFragpipeTMT.
}
