% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaFragpipeTMTdiGly.R
\name{limmaFragpipeTMTdiGly}
\alias{limmaFragpipeTMTdiGly}
\title{Imports normalized TMT diGly data from NormalyzerDE and performs limma contrasts.}
\usage{
limmaFragpipeTMTdiGly(
  inputPath,
  jobname,
  outputPath,
  replicateFilter = TRUE,
  proteinInput = TRUE,
  groups = FALSE,
  force = FALSE,
  exclude = NULL
)
}
\arguments{
\item{inputPath}{Input path where "metadata.csv", "contrast.csv", "abundance_single-site_MD.tsv" and "abundance_protein_MD.tsv" are located.}

\item{jobname}{Specifies the folder name for NormalyzerDE, for convenience I always use the PROT-XXXX.}

\item{replicateFilter}{Defaults to TRUE. Filters proteins that are not present in at least 2 replicates.}

\item{proteinInput}{Defaults to TRUE. Includes protein input for limma contrasts.}

\item{groups}{Defaults to FALSE. If TRUE, will use the type column in metadata.csv to search unique proteins based on those groupings.}

\item{force}{Defaults to FALSE. Forces saving the results table in case it already exists.}

\item{exclude}{Defaults to NULL. Defines a character vector of columns to be excluded from the pipeline.}

\item{outputpath}{Output path to save all the outputs.}
}
\value{
Generates the limma contrasts and saves them in the outputPath.
}
\description{
This function takes an inputPath with "metadata.csv", "contrast.csv",
"abundance_single-site_MD.tsv" and normalized protein abundances, and the chosen normalization
method from the NormalizerDE output folder, then normalizes each site with its paired input
sample, lastly performs limma differential gene expression
and saves all outputs in the ouputPath. jobname should be the same used from
importFragpipeTMTdiGly.
}
