% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStages.R
\name{plotStages}
\alias{plotStages}
\title{Plot stackplots from zebrafish staging}
\usage{
plotStages(inputDir, outputDir, p = 0.6, ctr, timepoints = c(2, 4, 6))
}
\arguments{
\item{inputDir}{Input directory where "classification_cluster.csv" and "sample_info.csv" are, with these names in the file name.}

\item{outputDir}{Output directory where plots and intermediate tables will be saved.}

\item{p}{Probability cutoff for considering the prediction correct. Default is 0.6}

\item{ctr}{Character vector with the names of your control conditions. It should match the column "condition" from "sample_info.csv".}

\item{timepoints}{Numeric vector with which time points you want to plot, and should match the values in column "time" from "sample_info.csv". Defaults to "c(2, 4, 6)"}
}
\value{
Plots and tables saved in the output folder path.
}
\description{
This function plots stackplots with zebrafish staging data and performs Fisher's exact test.
Sample tables used as input can be found inside "/inst/extdata/Staging_examples/".
}
