% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportTMT.R
\name{reportTMT}
\alias{reportTMT}
\title{Generates an html report based on outputs from limmaFragpipeTMT function}
\usage{
reportTMT(path, jobname)
}
\arguments{
\item{path}{Main path of the project, usually as PROT-XXXX and at the top level.}

\item{jobname}{Specifies the project name, for convenience I always use the PROT-XXXX.}
}
\value{
Generates a report with all standard analyses for TMT data.
}
\description{
This function takes the output from limmaFragpipeTMT function and generates a report with
all results and tables.
}
