% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportTMTdiGly.R
\name{reportTMTdiGly}
\alias{reportTMTdiGly}
\title{Generates an html report based on outputs from limmaFragpipeTMTdiGly function}
\usage{
reportTMTdiGly(path, jobname)
}
\arguments{
\item{path}{Main path of the project, usually as PROT-XXXX and at the top level.}

\item{jobname}{Specifies the project name, for convenience I always use the PROT-XXXX.}
}
\value{
Generates a report with all standard analyses for TMT diGly data.
}
\description{
This function takes the output from limmaFragpipeTMTdiGly function and generates a report with
all results and tables.
}
