% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverseAAseq.R
\name{reverseAAseq}
\alias{reverseAAseq}
\title{Reverses an aminoacid sequence in FASTA format to use as rev_ decoy.}
\usage{
reverseAAseq(sequence)
}
\arguments{
\item{sequence}{character vector of your aminoacid sequence.}
}
\value{
Generates a character vector with your reversed sequence.
}
\description{
This function takes a single character vector of an aminoacid sequence and reverses
the amino acid order to use a rev_ decoy for proteomics searches. It will also take out the
M from the end and put it as the first character.
}
