% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribosomeProteinsZf.R
\docType{data}
\name{ribosomeProteinsZf}
\alias{ribosomeProteinsZf}
\title{Ribosome subunits protein IDs from UNIPROT from zebrafish}
\format{
A data frame with 107 rows and 10 columns:
\describe{
\item{ProteinID}{Accession ID from uniprot}
\item{Reviewed}{If protein ID is from reviewed list or not}
\item{Entry Name}{Uniprot entry name}
\item{Protein names}{Long format of protein name}
\item{Gene Names}{Short protein names assosicated with ID, might have multiple}
\item{Organism}{Which organism it is from}
\item{Length}{Aminoacid length of protein}
\item{Ensembl}{Ensembl transcript ID mapped to the protein ID, might have multiple}
\item{ProteinName}{Gene names taken from the first name of Gene Names column, should be unique}
...
}
}
\source{
Downloaded from Uniprot on 2025/09/15.
}
\usage{
ribosomeProteinsZf
}
\description{
Dataframe with translation related IDs downloaded from uniprot and GO term GO:00058 on 2025/09/15
and filtered for gene names that start with rps or rpl ("^rps|^rpl").
}
\keyword{datasets}
