% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silacDIALabelingEfficiency.R
\name{silacDIALabelingEfficiency}
\alias{silacDIALabelingEfficiency}
\title{Imports SILAC PSM tables from Fragpipe-DIA mode and calculates labeling efficiency}
\usage{
silacDIALabelingEfficiency(inputPath, outputPath)
}
\arguments{
\item{inputPath}{Input path directing to 3_fragpipe/results.}

\item{outputpath}{Output path to save all the outputs.}
}
\value{
Generates the labeling efficiency stats and saves QC plots in outputPath.
}
\description{
This function takes an inputPath as the 3_fragpipe/results folder and reads all "psm.tsv" to
calculate SILAC labeling efficiency. It outputs a dataframe with all stats and saves a histogram
and a scatter plot for each sample found. NOTE: The ratio plots were not updated for the DIA input
files, still needs to be implemented. However, labeling efficiency can be calculated with this function.
}
