% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlopeCalc.R
\name{SlopeCalc}
\alias{SlopeCalc}
\title{Calculates linear regression for proteasome activity assay in batch}
\usage{
SlopeCalc(df, cutoff, conditions, types = "normal", replicates)
}
\arguments{
\item{df}{Melted dataframe input with "conditions", "type", "replicates", "time" and values.}

\item{cutoff}{Specifies the time cutoff so that calculation is only done in the exponential curve.}

\item{conditions}{Character vector with all conditions in the dataframe. Can be used as df$conditions.}

\item{types}{Character vector with all types in the dataframe. Used to differentiate between normal conditions or conditions where samples were treated with MG132. Default is "normal".}

\item{replicates}{Character vector with all replicates in the dataframe. Can be used as df$replicates}
}
\value{
Generates a dataframe with all slope values from lm().
}
\description{
This function takes a dataframe with different conditions and replicates from an enzymatic assay
like proteasome activity assay where you have signal that increases over time.
}
\details{
Note: You should plot the curves before using this to define the cutoff time and use only the
exponential time of the curve.
}
