% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stages.R
\docType{data}
\name{Stages}
\alias{Stages}
\title{Stages}
\format{
A data frame with 5 rows and 2 columns:
\describe{
\item{class}{Number related to class}
\item{stage}{Human readable stage}
...
}
}
\source{
Generated from classifier
}
\usage{
Stages
}
\description{
Dataframe with stages data for plotStages().
}
\keyword{datasets}
