% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtractContrastWide.R
\name{subtractContrastsWide}
\alias{subtractContrastsWide}
\title{Create site-ratios based on specific contrast tables for Ube3-APA package}
\usage{
subtractContrastsWide(df, contrastTable, diGly = TRUE)
}
\arguments{
\item{df}{Dataframe with abundances.}

\item{contrastTable}{Contrast table dataframe to be used.}

\item{diGly}{Defaults to TRUE. Selects "position" instead of just "id" from main dataframe.}
}
\value{
Generates a dataframe with site-ratios based on the contrast table.
}
\description{
This function takes a dataframe of protein abundances and generates site-ratios
based on a contrast table for Ube3-APA package.
}
