% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translationProteins.R
\docType{data}
\name{translationProteins}
\alias{translationProteins}
\title{Translation related protein IDs from UNIPROT}
\format{
A data frame with 534 rows and 9 columns:
\describe{
\item{ProteinID}{Accession ID from uniprot}
\item{Reviewed}{If protein ID is from reviewed list or not}
\item{Entry Name}{Uniprot entry name}
\item{Protein names}{Long format of protein name}
\item{Gene Names}{Short protein names assosicated with ID, might have multiple}
\item{Organism}{Which organism it is from}
\item{Length}{Aminoacid length of protein}
\item{Name}{Gene names taken from the first name of Gene Names column, should be unique}
...
}
}
\source{
Downloaded from Uniprot on 2025/09/15.
}
\usage{
translationProteins
}
\description{
Dataframe with translation related IDs downloaded from uniprot and GO term GO:00064 on 2025/09/15.
}
\keyword{datasets}
