% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniprotEnsemblMap.R
\docType{data}
\name{UniprotEnsemblMap}
\alias{UniprotEnsemblMap}
\title{Mapped UNIPROT IDs to Ensemble Gene IDs}
\format{
A data frame with 18675 rows and 3 columns:
\describe{
\item{Accession}{Accession ID from uniprot}
\item{ID}{ID}
\item{GeneID}{Ensembl GeneID}
...
}
}
\source{
Generated with biomaRt.
}
\usage{
UniprotEnsemblMap
}
\description{
Dataframe with mapped uniprot protien IDs to Ensemble gene IDs. This was generated with biomaRt on
2024/11/20.
}
\keyword{datasets}
