% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importDIAquickomics.R
\name{importDIAquickomics}
\alias{importDIAquickomics}
\title{Import DIA tables from ProteoVista quickomics input files}
\usage{
importDIAquickomics(inputPath, outputPath)
}
\arguments{
\item{inputPath}{Path where quickomics files are and should include msqrob_dea_results.csv, quickomics_expression_data.csv, quickomics_gene_protein_table.csv and quickomics_sample_metadata.csv}

\item{outputPath}{Path to where you want to save the output tables}
}
\value{
Assigns the new dataframes as "dea" and "abundance" in the global environment, while also saving the tables as .csv in the outputPath.
}
\description{
This function takes a path to the ProteoVista quickomics input files and cleans the tables so that
gene names are nicely after the protein ID, while also saving the output into the outputPath.
}
