% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaFragpipeTMT.R
\name{limmaFragpipeTMT}
\alias{limmaFragpipeTMT}
\title{Imports normalized TMT data from NormalyzerDE and performs limma contrasts.}
\usage{
limmaFragpipeTMT(inputPath, jobname, method = "quantile", outputPath)
}
\arguments{
\item{inputPath}{Input path where "metadata.csv", "contrast.csv" and "abundance_protein_MD.tsv" are located.}

\item{jobname}{Specifies the folder name for NormalyzerDE, for convenience I always use the PROT-XXXX.}

\item{method}{Specifies the chosen normalization method, defaults to "quantile". Options are: "cycloess", "gi", "log2", "mean", "median", "quantile", "rlr"}

\item{outputpath}{Output path to save all the outputs.}
}
\value{
Generates the limma contrasts and saves them in the outputPath.
}
\description{
This function takes an inputPath with "metadata.csv", "contrast.csv" and
"abundance_protein_MD.tsv", and the chosen normalization method from the
NormalizerDE output folder, then perform limma differential gene expression
and saves all outputs in the ouputPath. jobname should be the same used from
importFragpipeTMT.
}
