% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processDiannMSdap.R
\name{processDiannMSdap}
\alias{processDiannMSdap}
\title{Generates post-processing R script}
\usage{
processDiannMSdap(inputPath, outputPath)
}
\arguments{
\item{inputPath}{Path where input files report.tsv (diann), sample_metadata.xlsx, FASTA used in the search and contrast.csv are located.}

\item{outputPath}{Path where results will be saved.}
}
\value{
Creates all scripts to import DIA from diann to msdap and post-process.
}
\description{
This function generates an R script to run post-processing on msdap output.
It is a helper function for the processDiannMSdap.R function.
}
