% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMTlabelEfficiency.R
\name{TMTlabelEfficiency}
\alias{TMTlabelEfficiency}
\title{Calculates labeling efficiency from TMT files}
\usage{
TMTlabelEfficiency(inputPath)
}
\arguments{
\item{inputPath}{Input path where "psm.tsv" files are located.}
}
\value{
Returns a dataframe with calculations per subfolder found.
}
\description{
This function takes an inputPath from FragPipe results where several subfolders
contain the psm.tsv files generated and calculates labeling efficiency by counting
how many PSMs contain the TMT modification divided by the total number of PSMs.
}
